aruha-403 refactoring and removing unused code
The validation feature used to contain a lot of code that remained from
early proof of concepts. We are removing it since it was not used and
made things complicated.

We are also extracting one new class to encapsulate the loading of an
effective json schema.
