+++ src/main/java/org/zalando/nakadi/config/ValidatorConfig.java
@@ -14,8 +14,8 @@ import org.zalando.nakadi.domain.Version;
 import org.zalando.nakadi.validation.EventTypeOptionsValidator;
 import org.zalando.nakadi.validation.SchemaEvolutionService;
 import org.zalando.nakadi.validation.schema.CompatibilityModeChangeConstraint;
-import org.zalando.nakadi.validation.schema.FixedSchemaChangeConstraint;
 import org.zalando.nakadi.validation.schema.EnrichmentStrategyConstraint;
+import org.zalando.nakadi.validation.schema.FixedSchemaChangeConstraint;
 import org.zalando.nakadi.validation.schema.PartitionKeyFieldsConstraint;
 import org.zalando.nakadi.validation.schema.PartitionStrategyConstraint;
 import org.zalando.nakadi.validation.schema.SchemaDiff;
