+++ src/main/java/org/zalando/nakadi/domain/ValidationStrategyConfiguration.java
@@ -1,10 +1,7 @@
 package org.zalando.nakadi.domain;
 
-import org.json.JSONObject;
-
 public class ValidationStrategyConfiguration {
     private String strategyName;
-    private JSONObject additionalConfiguration;
 
     public String getStrategyName() {
         return strategyName;
@@ -12,13 +9,5 @@ public class ValidationStrategyConfiguration {
 
     public void setStrategyName(final String strategyName) {
         this.strategyName = strategyName;
-    }
-
-    public JSONObject getAdditionalConfiguration() {
-        return additionalConfiguration;
-    }
-
-    public void setAdditionalConfiguration(final JSONObject additionalConfiguration) {
-        this.additionalConfiguration = additionalConfiguration;
     }
 }
